<fix> replace labeled_dataloader with dkt.dataloader
